+++ src/Kernel/Console/SolitalCommands.php
@@ -5,34 +5,6 @@ namespace Solital\Core\Kernel\Console;
 use Solital\Core\Console\Interface\ExtendCommandsInterface;
 use Solital\Core\Console\Output\{ColorsEnum, ConsoleOutput};
 use Solital\Core\Kernel\{Application, DebugCore};
-use Solital\Core\Kernel\Console\Commands\{
-    AppStatus,
-    DumpDatabase,
-    GenerateConfigFiles,
-    GenerateHash,
-    HandleCache,
-    HandleCourse,
-    ListDatabase,
-    MakeController,
-    MakeAuth,
-    MakeBoot,
-    MakeCommand,
-    MakeMiddleware,
-    MakeMigrations,
-    MakeModel,
-    MakeQueue,
-    MakeRouter,
-    MakeSchedule,
-    MakeSeeder,
-    MakeValidator,
-    Migrations,
-    Queues,
-    Scanner,
-    Schedule,
-    Seeders,
-    Server,
-    Version
-};
 
 class SolitalCommands implements ExtendCommandsInterface
 {
@@ -44,7 +16,7 @@ class SolitalCommands implements ExtendCommandsInterface
     public function __construct()
     {
         if (DebugCore::isCoreDebugEnabled()) {
-            ConsoleOutput::banner("DEBUG ENABLED", ColorsEnum::BG_YELLOW)->print()->break();
+            ConsoleOutput::banner("DEBUG ENABLED", ColorsEnum::BG_YELLOW, 20)->print()->break();
         }
 
         Application::getInstance();
@@ -53,40 +25,14 @@ class SolitalCommands implements ExtendCommandsInterface
     /**
      * @var array
      */
-    protected array $command_class = [
-        AppStatus::class,
-        DumpDatabase::class,
-        GenerateConfigFiles::class,
-        GenerateHash::class,
-        HandleCache::class,
-        HandleCourse::class,
-        ListDatabase::class,
-        MakeController::class,
-        MakeBoot::class,
-        MakeAuth::class,
-        MakeCommand::class,
-        MakeMiddleware::class,
-        MakeMigrations::class,
-        MakeModel::class,
-        MakeQueue::class,
-        MakeRouter::class,
-        MakeSchedule::class,
-        MakeSeeder::class,
-        MakeValidator::class,
-        Migrations::class,
-        Queues::class,
-        Scanner::class,
-        Schedule::class,
-        Seeders::class,
-        Server::class,
-        Version::class
-    ];
+    protected array $command_class = [];
 
     /**
      * @return array
      */
     public function getCommandClass(): array
     {
+        $this->command_class = $this->getCommandsFiles();
         return $this->command_class;
     }
 
@@ -96,5 +42,33 @@ class SolitalCommands implements ExtendCommandsInterface
     public function getTypeCommands(): string
     {
         return $this->type_commands;
+    }
+
+    /**
+     * @return array
+     */
+    private static function getCommandsFiles(): array
+    {
+        $files = [];
+        $classes = [];
+        $cmd_dir = __DIR__ . DIRECTORY_SEPARATOR . "Commands" . DIRECTORY_SEPARATOR;
+        $handle = Application::provider("handler-file");
+
+        if (is_dir($cmd_dir)) {
+            $all_files = $handle->folder($cmd_dir)->files();
+
+            if (!is_null($all_files)) {
+                foreach ($all_files as $file) {
+                    $classes[] = basename($file, ".php");
+                }
+
+                foreach ($classes as $class) {
+                    $files[] = "\Solital\Core\Kernel\Console\Commands\\" . $class;
+                }
+            }
+        }
+
+        clearstatcache();
+        return $files;
     }
 }