+++ src/Resource/Str/Trait/StrStaticTrait.php
@@ -6,8 +6,6 @@ use Solital\Core\Resource\Str\Exceptions\InvalidStringArgumentException;
 
 trait StrStaticTrait
 {
-    use StrMultibyteTrait;
-
     /**
      * Does current string contain a subtring?
      * [!!] Not chainable
@@ -19,10 +17,7 @@ trait StrStaticTrait
      */
     public static function contains(string $haystack, string $needle): bool
     {
-        if (self::isBinary($haystack)) {
-            return self::mbContains($haystack, $needle);
-        }
-
+        if (self::isBinary($haystack)) return mb_str_contains($haystack, $needle);
         return str_contains($haystack, $needle);
     }
 
@@ -43,13 +38,13 @@ trait StrStaticTrait
      *                      - Mode 0, 1, or 2: returns an array with byte-values as keys and frequencies as values.
      *                      - Mode 3 or 4: returns a string with unique characters or unused characters.
      *
-     * @throws ValueError if the mode parameter is not between 0 and 4 (inclusive)
+     * @throws \ValueError if the mode parameter is not between 0 and 4 (inclusive)
      */
     public static function countChars(string $string, int $mode = 0): mixed
     {
-        return self::mbCountChars($string, $mode);
+        return mb_count_chars($string, $mode);
     }
-    
+
     /**
      * Binary safe string comparison
      * [!!] Not chainable
@@ -69,10 +64,7 @@ trait StrStaticTrait
      */
     public static function compare(string $string1, string $string2, ?string $encoding = null): int
     {
-        if (null === $encoding) {
-            $encoding = mb_internal_encoding();
-        }
-
+        if (null === $encoding) $encoding = mb_internal_encoding();
         return strcmp(mb_strtoupper($string1, $encoding), mb_strtoupper($string2, $encoding));
     }
 
@@ -93,10 +85,7 @@ trait StrStaticTrait
      */
     public static function compareInsensitive(string $string1, string $string2, ?string $encoding = null): int
     {
-        if (null === $encoding) {
-            $encoding = mb_internal_encoding();
-        }
-
+        if (null === $encoding) $encoding = mb_internal_encoding();
         return strcasecmp(mb_strtoupper($string1, $encoding), mb_strtoupper($string2, $encoding));
     }
 
@@ -117,17 +106,9 @@ trait StrStaticTrait
         string $encoding = 'UTF-8'
     ): bool {
         $validNeedle = self::checkString($needle);
-
-        if (0 === mb_strlen($validNeedle, $encoding)) {
-            return true;
-        }
-
+        if (0 === mb_strlen($validNeedle, $encoding)) return true;
         $haystackEnd = mb_substr($haystack, mb_strlen($validNeedle, $encoding) * -1, encoding: $encoding);
-
-        if ($caseInsensitive) {
-            return 0 === mb_stripos($haystackEnd, $validNeedle, 0, $encoding);
-        }
-
+        if ($caseInsensitive) return 0 === mb_stripos($haystackEnd, $validNeedle, 0, $encoding);
         return 0 === mb_strpos($haystackEnd, $validNeedle, 0, $encoding);
     }
 
@@ -152,23 +133,16 @@ trait StrStaticTrait
                 if (is_array($str)) {
                     $pos = self::position($haystack, $str);
                 } else {
-                    if (self::isBinary($haystack)) {
-                        $pos = mb_strpos($haystack, $str, $offset, $encoding);
-                    } else {
+                    (self::isBinary($haystack)) ?
+                        $pos = mb_strpos($haystack, $str, $offset, $encoding) :
                         $pos = strpos($haystack, $str, $offset);
-                    }
                 }
 
-                if ($pos !== FALSE) {
-                    return $pos;
-                }
+                if ($pos !== FALSE) return $pos;
             }
         }
 
-        if (self::isBinary($haystack)) {
-            return mb_strpos($haystack, $needles, $offset, $encoding);
-        }
-
+        if (self::isBinary($haystack)) return mb_strpos($haystack, $needles, $offset, $encoding);
         return strpos($haystack, $needles);
     }
 
@@ -189,15 +163,8 @@ trait StrStaticTrait
         string $encoding = 'UTF-8'
     ): bool {
         $validNeedle = self::checkString($needle);
-
-        if (0 === mb_strlen($validNeedle, $encoding)) {
-            return true;
-        }
-
-        if ($caseInsensitive) {
-            return 0 === mb_stripos($haystack, $validNeedle, 0, $encoding);
-        }
-
+        if (0 === mb_strlen($validNeedle, $encoding)) return true;
+        if ($caseInsensitive) return 0 === mb_stripos($haystack, $validNeedle, 0, $encoding);
         return 0 === mb_strpos($haystack, $validNeedle, 0, $encoding);
     }
 
@@ -210,10 +177,7 @@ trait StrStaticTrait
      */
     public static function uniqueChars(string $string): string
     {
-        if (self::isBinary($string)) {
-            return self::mbCountChars($string, 3);
-        }
-
+        if (self::isBinary($string)) return mb_count_chars($string, 3);
         return count_chars($string, 3);
     }
 
@@ -227,10 +191,7 @@ trait StrStaticTrait
      */
     public static function wordCount(string $string, ?string $charlist = null): array|int
     {
-        if (self::isBinary($string)) {
-            return self::mbWordCount($string, 0, $charlist);
-        }
-
+        if (self::isBinary($string)) return mb_str_word_count($string, 0, $charlist);
         return str_word_count($string, 0, $charlist);
     }
 
@@ -244,10 +205,7 @@ trait StrStaticTrait
      */
     public static function words(string $string, ?string $charlist = null): array
     {
-        if (self::isBinary($string)) {
-            return self::mbWordCount($string, 2, $charlist);
-        }
-
+        if (self::isBinary($string)) return mb_str_word_count($string, 2, $charlist);
         return str_word_count($string, 2, $charlist);
     }
 
@@ -268,10 +226,7 @@ trait StrStaticTrait
      */
     protected static function checkString(mixed $value): string
     {
-        if (is_scalar($value) || (is_object($value) && method_exists($value, '__toString'))) {
-            return strval($value);
-        }
-
+        if (is_scalar($value) || (is_object($value) && method_exists($value, '__toString'))) return strval($value);
         throw new InvalidStringArgumentException();
     }
 }