+++ src/Schedule/Attribute/TimeInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Solital\Core\Schedule\Attribute;
+
+interface TimeInterface
+{
+    public function getAttributeName(): string;
+    public function getArgs(): array;
+}