+++ tests/Resource/Str/StrMaskTest.php
@@ -1,7 +1,7 @@
 <?php
 
 use PHPUnit\Framework\TestCase;
-use Solital\Core\Resource\Str\StrMask;
+use Solital\Core\Resource\Str\Str;
 
 class StrMaskTest extends TestCase
 {
@@ -11,12 +11,12 @@ class StrMaskTest extends TestCase
         // Basic
         $this->assertEquals(
             "4 1 5",
-            StrMask::apply("415", "0 0 0")
+            Str::mask("415", "0 0 0")
         );
         // Complex
         $this->assertEquals(
             "4 1 5",
-            StrMask::apply("4ab1-5bh", "0 0 0")
+            Str::mask("4ab1-5bh", "0 0 0")
         );
     }
 
@@ -25,23 +25,23 @@ class StrMaskTest extends TestCase
         // Basic
         $this->assertEquals(
             "4 1 5",
-            StrMask::apply("415", "9 9 9")
+            Str::mask("415", "9 9 9")
         );
         // Complex 1
         $this->assertEquals(
             "4 1 5",
-            StrMask::apply("4ab15bh", "9 9 9")
+            Str::mask("4ab15bh", "9 9 9")
         );
         // Complex 2
         // TODO: To resolve, do not pass
         /*$this->assertEquals(
       "4 1 5",
-      StrMask::apply("4ab1.5bh", "9 9 9")
+      Str::mask("4ab1.5bh", "9 9 9")
     );*/
         // Complex 3
         $this->assertEquals(
             "4 12 553 1",
-            StrMask::apply("4.12.553d1", "099 099 099 099")
+            Str::mask("4.12.553d1", "099 099 099 099")
         );
     }
 
@@ -50,12 +50,12 @@ class StrMaskTest extends TestCase
         // Basic
         $this->assertEquals(
             "4 aB 5",
-            StrMask::apply("4aB5", "A AA A")
+            Str::mask("4aB5", "A AA A")
         );
         // Complex
         $this->assertEquals(
             "4 ab 15",
-            StrMask::apply("4ab1-5bh", "A AA AA")
+            Str::mask("4ab1-5bh", "A AA AA")
         );
     }
 
@@ -64,12 +64,12 @@ class StrMaskTest extends TestCase
         // Basic
         $this->assertEquals(
             "a B",
-            StrMask::apply("4aB5", "S SS S")
+            Str::mask("4aB5", "S SS S")
         );
         // Complex
         $this->assertEquals(
             "a bB h",
-            StrMask::apply("4ab1-5Bh", "S SS S")
+            Str::mask("4ab1-5Bh", "S SS S")
         );
     }
 
@@ -79,17 +79,17 @@ class StrMaskTest extends TestCase
         // Basic
         $this->assertEquals(
             "25.5639",
-            StrMask::apply("The result is 25.5639%", "percent")
+            Str::mask("The result is 25.5639%", "percent")
         );
         // With precision 2 decimals
         $this->assertEquals(
             "25.56",
-            StrMask::apply("The result is 25.5639%", "percent.2")
+            Str::mask("The result is 25.5639%", "percent.2")
         );
         // With precision 0 decimal
         $this->assertEquals(
             "25",
-            StrMask::apply("The result is 25.5639%", "percent.0")
+            Str::mask("The result is 25.5639%", "percent.0")
         );
     }
 
@@ -98,17 +98,17 @@ class StrMaskTest extends TestCase
         // Basic
         $this->assertEquals(
             "1 963 725,5630",   // Function use float and precision is lost
-            StrMask::apply("The result is 1963725.5639", "separator")
+            Str::mask("The result is 1963725.5639", "separator")
         );
         // With precision 2 decimals
         $this->assertEquals(
             "1 963 725,56",
-            StrMask::apply("The result is 1963725.5639", "separator.2")
+            Str::mask("The result is 1963725.5639", "separator.2")
         );
         // With precision 0 decimal
         $this->assertEquals(
             "1 963 726",
-            StrMask::apply("The result is 1963725.5639", "separator.0")
+            Str::mask("The result is 1963725.5639", "separator.0")
         );
     }
 
@@ -117,17 +117,17 @@ class StrMaskTest extends TestCase
         // Basic
         $this->assertEquals(
             "1.963.725,5639",
-            StrMask::apply("The result is 1963725.5639", "dot_separator")
+            Str::mask("The result is 1963725.5639", "dot_separator")
         );
         // With precision 2 decimals
         $this->assertEquals(
             "1.963.725,56",
-            StrMask::apply("The result is 1963725.5639", "dot_separator.2")
+            Str::mask("The result is 1963725.5639", "dot_separator.2")
         );
         // With precision 0 decimal
         $this->assertEquals(
             "1.963.726",
-            StrMask::apply("The result is 1963725.5639", "dot_separator.0")
+            Str::mask("The result is 1963725.5639", "dot_separator.0")
         );
     }
 
@@ -136,17 +136,17 @@ class StrMaskTest extends TestCase
         // Basic
         $this->assertEquals(
             "1,963,725.5639",
-            StrMask::apply("The result is 1963725.5639", "comma_separator")
+            Str::mask("The result is 1963725.5639", "comma_separator")
         );
         // With precision 2 decimals
         $this->assertEquals(
             "1,963,725.56",
-            StrMask::apply("The result is 1963725.5639", "comma_separator.2")
+            Str::mask("The result is 1963725.5639", "comma_separator.2")
         );
         // With precision 0 decimal
         $this->assertEquals(
             "1,963,726",
-            StrMask::apply("The result is 1963725.5639", "comma_separator.0")
+            Str::mask("The result is 1963725.5639", "comma_separator.0")
         );
     }
 
@@ -155,7 +155,7 @@ class StrMaskTest extends TestCase
     {
         $this->assertEquals(
             "The result is 48",
-            StrMask::apply("If you set 4 next to 8, what is the result?", "00", array(
+            Str::mask("If you set 4 next to 8, what is the result?", "00", array(
                 "prefix" => "The result is "
             ))
         );
@@ -164,7 +164,7 @@ class StrMaskTest extends TestCase
     {
         $this->assertEquals(
             "48 is the result!",
-            StrMask::apply("If you set 4 next to 8, what is the result?", "00", array(
+            Str::mask("If you set 4 next to 8, what is the result?", "00", array(
                 "suffix" => " is the result!"
             ))
         );